delete `spec` folder
