--- spec/index.js
@@ -1,2 +1,0 @@
-import './mock.js';
-import './utils.test.js';
