+++ test/setup/jsdom.js
@@ -22,5 +22,3 @@ global.L = {
         hasClass: () => {}
     }
 };
-
-
